+++ frontend/src/components/common/visualizer/elements/base.jsx
