+++ frontend/src/components/common/visualizer/elements/verticalList.jsx
