+++ frontend/src/components/jobs/result/JobOverview.jsx
@@ -25,7 +25,7 @@ import {
   ReportedPluginTooltip,
 } from "./utils/reportedPlugins";
 import { StatusIcon } from "../../common/icon/StatusIcon";
-import VisualizerReport from "./visualizer/visualizer";
+import VisualizerReport from "../../common/visualizer/visualizer";
 import { JobFinalStatuses } from "../../../constants/jobConst";
 import { PluginStatuses } from "../../../constants/pluginConst";
 import { JobResultSections } from "../../../constants/miscConst";
