+++ frontend/tests/components/jobs/result/visualizer/elements/base.test.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import "@testing-library/jest-dom";
 import { render, screen, waitFor } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
-import { BaseVisualizer } from "../../../../../../src/components/jobs/result/visualizer/elements/base";
+import { BaseVisualizer } from "../../../../../../src/components/common/visualizer/elements/base";
 import { getIcon } from "../../../../../../src/components/common/icon/icons";
 
 // mock useLocation
