+++ frontend/tests/components/jobs/result/visualizer/elements/download.test.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import "@testing-library/jest-dom";
 import { render, screen, waitFor } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
-import { DownloadVisualizer } from "../../../../../../src/components/jobs/result/visualizer/elements/download";
+import { DownloadVisualizer } from "../../../../../../src/components/common/visualizer/elements/download";
 
 // mock useLocation
 jest.mock("react-router-dom", () => ({
