+++ frontend/tests/components/jobs/result/visualizer/elements/verticalList.test.jsx
@@ -1,9 +1,9 @@
 import React from "react";
 import "@testing-library/jest-dom";
 import { render, screen } from "@testing-library/react";
-import { BaseVisualizer } from "../../../../../../src/components/jobs/result/visualizer/elements/base";
-import { VerticalListVisualizer } from "../../../../../../src/components/jobs/result/visualizer/elements/verticalList";
-import { HorizontalListVisualizer } from "../../../../../../src/components/jobs/result/visualizer/elements/horizontalList";
+import { BaseVisualizer } from "../../../../../../src/components/common/visualizer/elements/base";
+import { VerticalListVisualizer } from "../../../../../../src/components/common/visualizer/elements/verticalList";
+import { HorizontalListVisualizer } from "../../../../../../src/components/common/visualizer/elements/horizontalList";
 
 // mock useLocation
 jest.mock("react-router-dom", () => ({
